Backend/requirements.txt
django
django djangorestframework
django-cors-headers
Pillow
django-ckeditor
django-taggit
